requirements.txt
asgiref==3.8.1
Django==5.1.3
image==1.5.33
numpy==2.1.3
pillow==11.0.0
six==1.16.0
sqlparse==0.5.2
